requirements.txt
pandas==1.1.5
numpy==1.18.5
pandas-gbq==0.14.1
google-auth==1.32.1
google-cloud-storage==1.42.0
sklearn==0.0
pyarrow>=0.4